requirements.txt
fastapi==0.136.0
pillow==12.2.0
requests==2.33.1
vercel==0.5.8
